Move logic for canItemBeRemovedFromCart to wpcom-checkout package
