+++ packages/wpcom-checkout/src/index.ts
@@ -26,3 +26,4 @@ export * from './get-introductory-offer-interval-display';
 export * from './join-classes';
 export * from './checkout-line-items';
 export * from './get-country-postal-code-support';
+export * from './can-item-be-removed-from-cart';
